refactor(backend): extract roleservice from guildservice
Behavior-preserving extraction of role logic into dedicated RoleService.
Moved 5 methods (getGuildRoleOptions, getFullGuildRoles, createGuildRole,
updateGuildRole, deleteGuildRole) and related interfaces/validators.

GuildService retains thin one-line delegates for backward compatibility.
All callers (routes/roles.ts, RoleGroupService.ts) remain unchanged.
